telegram-telethon/pyproject.toml
[project]
name = "telegram-telethon"
version = "0.1.0"
description = "Full Telethon API wrapper with daemon mode and Claude Code integration"
requires-python = ">=3.10"
dependencies = [
    "telethon>=1.36.0",
    "pyyaml>=6.0",
    "rich>=13.0",           # CLI formatting
    "questionary>=2.0",      # Interactive prompts
    "aiofiles>=23.0",        # Async file I/O
    "httpx>=0.27.0",         # Groq API for transcription
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
]

[project.scripts]
tg = "telegram_telethon.cli:main"
tgd = "telegram_telethon.daemon.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (requires Telegram connection)",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src/telegram_telethon"]
omit = ["tests/*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
